planet/requirements.txt
requests
boto3
kubernetes
pulsar-client
planet
pycryptodome